Prepare for testing, add Dialogflow
Create a separate assistant component to allow using different recognizer
modules with the same assistant algorithm. This could be used for testing
with pre-recorded audio.

Initial version of command processing with DialogFlow. Once the keyword
was detected we start sending the context and the command to DialogFlow
for intent recognition.